requirements.txt
groq>=0.11.0
python-dotenv==1.0.0
flask==3.0.0
